Reorganize the DataManager class into its own file and add preprocess_svd method to it for LightCurveTrainer. Also removed now superfluous lines at the end of neuralnets.py.
