+++ src/fiesta/train/LightcurveTrainer.py
@@ -6,12 +6,14 @@ import numpy as np
 import jax
 import jax.numpy as jnp 
 from jaxtyping import Array, Float, Int
+from typing import Dict
 from fiesta.utils import MinMaxScalerJax
 from fiesta import utils
 from fiesta.utils import Filter
 from fiesta import conversions
 from fiesta.constants import days_to_seconds, c
 from fiesta import models_utilities
+from fiesta.train.DataManager import DataManager
 import fiesta.train.neuralnets as fiesta_nn
 
 import matplotlib.pyplot as plt
@@ -19,68 +21,52 @@ import pickle
 from typing import Callable
 import tqdm
 
-import afterglowpy as grb
+################
+# TRAINING API #
+################
 
-class SurrogateTrainer:
+class LightcurveTrainer:
     """Abstract class for training a collection of surrogate models per filter"""
-    
+   
     name: str
     outdir: str
     filters: list[Filter]
     parameter_names: list[str]
-    
-    preprocessing_metadata: dict[str, dict[str, float]]
-    
-    # TODO: why do we have so many datasets?
-    X: Float[Array, "n_batch n_input_surrogate"]
-    y: dict[str, Float[Array, "n_batch n_output_surrogate"]]
-    
-    X_raw: Float[Array, "n_batch n_params"]
-    y_raw: dict[str, Float[Array, "n_batch n_times"]]
-    
-    train_X: Float[Array, "n_batch n_params"]
-    train_y: dict[str, Float[Array, "n_batch n_times"]]
-    
-    val_X: Float[Array, "n_batch n_params"]
-    val_y: dict[str, Float[Array, "n_batch n_times"]]
-    
-    train_X_raw: Float[Array, "n_batch n_params"]
-    train_y_raw: dict[str, Float[Array, "n_batch n_times"]]
-    
-    val_X_raw: Float[Array, "n_batch n_params"]
-    val_y_raw: dict[str, Float[Array, "n_batch n_times"]]
-    
-    trained_states: dict[str, fiesta_nn.TrainState]
+
+    train_X: Float[Array, "n_train"]
+    train_y: Dict[str, Float[Array, "n"]]
+    val_X: Float[Array, "n_val"]
+    val_y: Dict[str, Float[Array, "n"]]
     
     def __init__(self, 
                  name: str,
                  outdir: str,
-                 save_raw_data: bool = False,
+                 plots_dir: str = None,
                  save_preprocessed_data: bool = False) -> None:
         
         self.name = name
-        self.outdir = outdir
         # Check if directories exists, otherwise, create:
+        self.outdir = outdir
         if not os.path.exists(self.outdir):
             os.makedirs(self.outdir)
-        
-        self.save_raw_data = save_raw_data
+        self.plots_dir = plots_dir
+        if self.plots_dir is not None and not os.path.exists(self.plots_dir):
+            os.makedirs(self.plots_dir)
+
         self.save_preprocessed_data = save_preprocessed_data
-        
+
         # To be loaded by child classes
         self.filters = None
         self.parameter_names = None
-        self.preprocessing_metadata = {}
         
-        self.X_raw = None
-        self.y_raw = None
-        
-        self.X = None
-        self.y = None
-        self.weights = None
+        self.train_X = None
+        self.train_y = None
+
+        self.val_X = None
+        self.val_y = None
 
     def __repr__(self) -> str:
-        return f"SurrogateTrainer(name={self.name})"
+        return f"LightcurveTrainer(name={self.name})"
     
     def preprocess(self):
         
@@ -105,7 +91,7 @@ class SurrogateTrainer:
     def fit(self,
             config: fiesta_nn.NeuralnetConfig = None,
             key: jax.random.PRNGKey = jax.random.PRNGKey(0),
-            verbose: bool = True):
+            verbose: bool = True) -> None:
         """
         The config controls which architecture is built and therefore should not be specified here.
         
@@ -156,76 +142,42 @@ class SurrogateTrainer:
         """
         Save the trained model and all the used metadata to the outdir.
         """
-        if not os.path.exists(self.outdir):
-            os.makedirs(self.outdir)
-        
+        # Save the metadata
         meta_filename = os.path.join(self.outdir, f"{self.name}_metadata.pkl")
 
-        # FIXME: this should not be in this class
-        if os.path.exists(meta_filename):
-            with open(meta_filename, "rb") as meta_file:
-                save = pickle.load(meta_file)
-            if not np.array_equal(save["times"], self.times): # check whether the metadata from previously trained filters agrees
-                raise Exception(f"The time array needs to coincide with the time array for previous filters: {save['times']}")
-            if not np.array_equal(save["parameter_names"], self.parameter_names):
-                 raise Exception(f"The parameters need to coincide with the parameters for previous filters: {save['parameter_names']}")
-        else:
-            save = {}
+        save = {}
 
         save["times"] = self.times
         save["parameter_names"] = self.parameter_names
-        # TODO: see if we can save the jet_type here somewhat more self-consistently
 
-        for filt in self.filters:
-            model = self.trained_states[filt.name]
-            fiesta_nn.save_model(model, self.config, out_name=self.outdir + f"{filt.name}.pkl")
-            save[filt.name] = self.preprocessing_metadata[filt.name]
-        
         with open(meta_filename, "wb") as meta_file:
             pickle.dump(save, meta_file)
         
-    def _save_raw_data(self):
-        print("Saving raw data . . .")
-        np.savez(os.path.join(self.outdir, "raw_data_training.npz"), X_raw=self.train_X_raw, **self.train_y_raw)
-        np.savez(os.path.join(self.outdir, "raw_data_validation.npz"), X_raw=self.val_X_raw, **self.val_y_raw)
-        print("Saving raw data . . . done")
+        # Save the NN
+        for filt in self.filters:
+            model = self.trained_states[filt.name]
+            model.save_model(outfile = os.path.join(self.outdir, f"{self.name}_{filt.name}.pkl"))
+            save[filt.name] = self.preprocessing_metadata[filt.name]
         
-    def _save_preprocessed_data(self):
+
+        
+    def _save_preprocessed_data(self) -> None:
         print("Saving preprocessed data . . .")
-        np.savez(os.path.join(self.outdir, "preprocessed_data_training.npz"), X=self.train_X, **self.train_y)
-        np.savez(os.path.join(self.outdir, "preprocessed_data_validation.npz"), X=self.val_X, **self.val_y)
+        np.savez(os.path.join(self.outdir, f"{self.name}_preprocessed_data.npz"), train_X=self.train_X, train_y = self.train_y, val_X = self.val_X, val_y = self.val_y)
         print("Saving preprocessed data . . . done")
     
-class SVDSurrogateTrainer(SurrogateTrainer):
-    
-    outdir: str 
-    svd_ncoeff: Int
-    tmin: Float
-    tmax: Float
-    dt: Float
-    times: Float[Array, "n_times"]
-    plots_dir: str
-    save_raw_data: bool
-    save_preprocessed_data: bool
+class SVDTrainer(LightcurveTrainer):
     
     def __init__(self,
                  name: str,
                  outdir: str,
-                 filters: list[str] = None,
-                 svd_ncoeff: Int = 10, 
-                 validation_fraction: Float = 0.2,
-                 tmin: Float = None,
-                 tmax: Float = None,
-                 dt: Float = None,
+                 filters: list[str],
+                 data_manager_args: dict,
+                 svd_ncoeff: Int = 10,
                  plots_dir: str = None,
-                 save_raw_data: bool = False,
-                 save_preprocessed_data: bool = False
-                 ):
-        
+                 save_preprocessed_data: bool = False) -> None:
         """
         Initialize the surrogate model trainer that uses an SVD. The initialization also takes care of reading data and preprocessing it, but does not automatically fit the model. Users may want to inspect the data before fitting the model.
-        
-        Note: currently, only models of Bulla type .dat files are supported
         
         Args:
             name (str): Name of the surrogate model. Will be used 
@@ -241,114 +193,41 @@ class SVDSurrogateTrainer(SurrogateTrainer):
             save_raw_data (bool, optional): If True, the raw data will be saved in the outdir. Defaults to False.
             save_preprocessed_data: If True, the preprocessed data (reduced, rescaled) will be saved in the outdir. Defaults to False.
         """
-        
-        super().__init__(name=name, outdir=outdir)
-        self.plots_dir = plots_dir
-        
-        if self.plots_dir is not None and not os.path.exists(self.plots_dir):
-            os.makedirs(self.plots_dir)
 
-        self.validation_fraction = validation_fraction
-            
-        self.svd_ncoeff = svd_ncoeff
-        self.tmin = tmin
-        self.tmax = tmax
-        self.dt = dt
-        self.save_raw_data = save_raw_data
-        self.save_preprocessed_data = save_preprocessed_data
+        super().__init__(name = name,
+                         outdir = outdir,
+                         plots_dir = plots_dir,
+                         save_preprocessed_data = save_preprocessed_data)
         
+        self.svd_ncoeff = svd_ncoeff
+        
+        self.data_manager = DataManager(**data_manager_args)
+        self.data_manager.print_file_info()
+        self.data_manager.pass_meta_data(self)
         self.load_filters(filters)
-        self.load_times()
-        self.load_parameter_names()
-            
-        self.load_raw_data()
 
         self.preprocess()
-    
-        if save_preprocessed_data:
+        if self.save_preprocessed_data:
             self._save_preprocessed_data()
-        if save_raw_data:
-            self._save_raw_data()
-        
-    def load_parameter_names(self):
-        raise NotImplementedError
-        
-    def load_times(self):
-        raise NotImplementedError
     
-    def load_filters(self, filters: list[str] = None):
-        raise NotImplementedError
-    
-    def load_raw_data(self):
-        raise NotImplementedError
+    def load_filters(self, filters):
+        self.filters = []
+        for filt in filters:
+            Filt = Filter(filt)
+            if Filt.nus[0] < self.nus[0] or Filt.nus[-1] > self.nus[-1]:
+                raise ValueError(f"Filter {filt} exceeds the frequency range of the training data.")
+            self.filters.append(Filt)
         
     def preprocess(self):
         """
-        Preprocess the data. This includes scaling the inputs and outputs, performing SVD decomposition, and saving the necessary metadata for later use.
+        Preprocessing method to get the SVD coefficients of the MinMaxScaler. This includes scaling the inputs and outputs, performing SVD decomposition, and saving the necessary metadata for later use.
         """
-        # Scale inputs
-        X_scaler = MinMaxScalerJax()
-        self.train_X = X_scaler.fit_transform(self.train_X_raw) # fit the scaler to the training data
-        self.val_X = X_scaler.transform(self.val_X_raw) # transform the val data
-      
-        # Scale outputs, do SVD and save into y
-        self.train_y = {filt.name: [] for filt in self.filters}
-        self.val_y = {filt.name: [] for filt in self.filters}
-        
-        print(f"Rescaling the training and validation data for filters {[filter.name for filter in self.filters]}")
-        for filt in tqdm.tqdm(self.filters):
-            
-            y_scaler = MinMaxScalerJax()
-
-            completely_problematic = np.where(np.all(np.isinf(self.train_y_raw[filt.name]), axis = 1))[0]
-            problematic = np.unique(np.where(np.isinf(self.train_y_raw[filt.name]))[0])
-
-            if len(problematic)!=0:
-                raise Exception(f"There were infs in the magnitudes for filter {filt.name}.")
-
-            data = y_scaler.fit_transform(self.train_y_raw[filt.name])
-            
-            # Do SVD decomposition on the training data
-            UA, _, VA = np.linalg.svd(data, full_matrices=True)
-            VA = VA.T
-
-            n, n = UA.shape
-            m, m = VA.shape
-            # This is taken over from NMMA
-            cAmat = np.zeros((self.svd_ncoeff, n))
-            cAvar = np.zeros((self.svd_ncoeff, n))
-            for i in range(n):
-                ErrorLevel = 1.0
-                cAmat[:, i] = np.dot(
-                    data[i, :], VA[:, : self.svd_ncoeff]
-                )
-                errors = ErrorLevel * np.ones_like(data[i, :])
-                cAvar[:, i] = np.diag(
-                    np.dot(
-                        VA[:, : self.svd_ncoeff].T,
-                        np.dot(np.diag(np.power(errors, 2.0)), VA[:, : self.svd_ncoeff]),
-                    )
-                )
-
-            self.train_y[filt.name] = cAmat.T # Transpose to get the shape (n_batch, n_svd_coeff)
-
-            # Do SVD decomposition on the validation data
-            val_data = y_scaler.transform(self.val_y_raw[filt.name])
-            cAmat = np.zeros((self.svd_ncoeff, self.n_val_data))
-            for i in range(self.n_val_data):
-                cAmat[:,i] = np.dot(
-                    val_data[i,:], VA[:, : self.svd_ncoeff]
-                )
-            
-            self.val_y[filt.name] = cAmat.T # Transpose to get the shape (n_val, n_svd_coeff)
-
-            #Save the scalers
-            self.preprocessing_metadata[filt.name] = {"VA": VA, "X_scaler_max": X_scaler.max_val, "X_scaler_min": X_scaler.min_val, "y_scaler": y_scaler, "svd_ncoeff": self.svd_ncoeff}
-
-        
-    def __repr__(self) -> str:
-        return f"SVDSurrogateTrainer(name={self.name}, lc_dir={self.lc_dir}, outdir={self.outdir}, filters={self.filters})"
-    
+        print(f"Decomposing training data to SVD coefficients.")
+        self.train_X, self.train_y, self.val_X, self.val_y, self.X_scaler, self.y_scaler = self.data_manager.preprocess_svd(self.svd_ncoeff, self.filters)
+        for key in self.train_y.keys():
+            if np.any(np.isnan(self.train_y[key])) or np.any(np.isnan(self.val_y[key])):
+                raise ValueError(f"Data preprocessing for {key} introduced nans. Check raw data for nans of infs or vanishing variance in a specific entry.")
+        print(f"Preprocessing data . . . done")
         
 class BullaSurrogateTrainer(SVDSurrogateTrainer):
     
@@ -484,239 +363,3 @@ class BullaSurrogateTrainer(SVDSurrogateTrainer):
         for filt in self.filters:
             self.train_y_raw[filt.name] = y_raw[filt.name][~mask]
             self.val_y_raw[filt.name] = y_raw[filt.name][mask]
-
-        
-# TODO: perhaps rename to *_1D, since it is only for 1D light curves, and we might want to get support for 2D by incorporating the frequencies... Unsure about the approach here
-
-class AfterglowpyTrainer(SVDSurrogateTrainer):
-    
-    parameter_grid: dict[str, list[Float]]
-    n_training_data: Int
-    fixed_parameters: dict[str, Float]
-    jet_type: Int
-    use_log_spacing: bool
-    
-    _times_afterglowpy: Float[Array, "n_times"]
-    nus: dict[str, Float]
-    
-    def __init__(self,
-                 name: str,
-                 outdir: str,
-                 filters: list[str],
-                 parameter_grid: dict[str, list[float]],
-                 weight_grids: dict = None,
-                 n_training_data: Int = 5000,
-                 jet_type: Int = -1,
-                 fixed_parameters: dict[str, Float] = {},
-                 tmin: Float = 0.1,
-                 tmax: Float = 1000,
-                 n_times: Int = 100,
-                 use_log_spacing: bool = True,
-                 validation_fraction: float = 0.2,
-                 plots_dir: str = None,
-                 svd_ncoeff: Int = 10,
-                 save_raw_data: bool = False,
-                 save_preprocessed_data: bool = False,
-                 remake_training_data = False,
-                 ):
-        """
-        Initialize the surrogate model trainer. The initialization also takes care of reading data and preprocessing it, but does not automatically fit the model. Users may want to inspect the data before fitting the model.
-        
-        Args:
-            name (str): Name given to the model
-            outdir (str): Output directory to save the trained model
-            parameter_grid (dict[str, list[Float]]): Dictionary containing the grid points for each parameter, i.e., the parameter values on which the surrogate will be trained. The keys should be the parameter names and the values should be a list..
-            jet_type (Int): Type of jet for the afterglowpy, -1 is tophat, 0 is Gaussian, 4 is PowerLaw
-            fixed_parameters (dict[str, Float]) : values of the afterglowpy parameters that should be kept fixed for the surrogate model
-            tmin (Float, optional): Minimum time in days of the light curve, all data before is discarded. Defaults to 0.1.
-            tmax (Float, optional): Maximum time in days of the light curve, all data after is discarded. Defaults to 1000.
-            n_times: number of time nodes for the training light curve data
-            use_log_spacing: bool : whether the time nodes of the training light curve data should be log10 spaced
-            validation_fraction (Float, optional): Fraction of the data to be used for validation. Defaults to 0.2.
-            plots_dir : str : outdir for the plots
-            svd_ncoeff: int : Number of SVD coefficients to use in data reduction during training. Defaults to 10.
-            save_raw_data (bool, optional): If True, the raw data will be saved in the outdir. Defaults to False.
-            save_preprocessed_data: If True, the preprocessed data (reduced, rescaled) will be saved in the outdir. Defaults to False.
-        """
-        
-
-        self.n_times = n_times
-        dt = (tmax - tmin) / n_times
-        self.parameter_grid = parameter_grid
-        self.weight_grids = weight_grids
-        if self.weight_grids is None:
-            self.weight_grids = {p: np.full_like(self.parameter_grid[p], 1/len(self.parameter_grid[p])) for p in self.parameter_grid.keys()}
-
-        self.fixed_parameters = fixed_parameters
-        self.use_log_spacing = use_log_spacing
-        
-        # Check jet type before saving
-        supported_jet_types = [-1, 0, 4]
-        if jet_type not in supported_jet_types:
-            raise ValueError(f"Jet type {jet_type} is not supported. Supported jet types are: {supported_jet_types}")
-        self.jet_type = jet_type
-        self.remake_training_data = remake_training_data
-
-        self.n_training_data = n_training_data
-        self.validation_fraction = validation_fraction
-        self.n_val_data = int(self.n_training_data * self.validation_fraction/(1-self.validation_fraction))
-            
-        super().__init__(name=name,
-                         outdir=outdir,
-                         filters=filters,
-                         svd_ncoeff=svd_ncoeff,
-                         validation_fraction=validation_fraction,
-                         tmin=tmin,
-                         tmax=tmax,
-                         dt=dt,
-                         plots_dir=plots_dir,
-                         save_raw_data=save_raw_data,
-                         save_preprocessed_data=save_preprocessed_data)
-       
-        
-    def load_filters(self, filters: list[str]):
-        self.filters = []
-        for filter in filters:
-            try:
-                self.filters.append(Filter(filter))
-            except:
-                raise Exception(f"Filter {filter} not available.")
-
-    def load_times(self):
-        if self.use_log_spacing:
-            times = np.logspace(np.log10(self.tmin), np.log10(self.tmax), num=self.n_times)
-        else:
-            times = np.linspace(self.tmin, self.tmax, num=self.n_times)
-        self.times = times
-        self._times_afterglowpy = self.times * days_to_seconds # afterglowpy takes seconds as input
-        
-    def load_parameter_names(self):
-        self.parameter_names = list(self.parameter_grid.keys())
-    
-    def load_raw_data(self):
-        data_files_exist =  os.path.exists(self.outdir+"/raw_data_training.npz") and os.path.exists(self.outdir+"/raw_data_validation.npz")
-        if data_files_exist and not self.remake_training_data:
-            self.train_X_raw, self.train_y_raw, self.val_X_raw, self.val_y_raw = self._read_files()
-        else:
-            self.create_raw_data()
-    
-    def create_raw_data(self):
-        """
-        Create a grid of training data with specified settings and generate the output files for them.
-
-        TODO: for now we train per filter, but best to change this!
-        """
-        # Create training data
-        nus = jnp.array([filt.nu for filt in self.filters])
-        X_raw = np.empty((self.n_training_data, len(self.parameter_names)))
-        y_raw = {filt.name: np.empty((self.n_training_data, len(self.times))) for filt in self.filters}
-
-        for j, key in enumerate(self.parameter_grid.keys()):
-            X_raw[:,j] = np.random.choice(self.parameter_grid[key], size = self.n_training_data, replace = True, p = self.weight_grids[key])
-
-
-        print(f"Creating the afterglowpy training dataset on grid with {self.n_training_data} points.")
-        for i in tqdm.tqdm(range(self.n_training_data)):
-            param_dict = dict(zip(self.parameter_names, X_raw[i]))
-            param_dict.update(self.fixed_parameters)
-            mJys = self._call_afterglowpy(param_dict, nus)
-            for k, filt in enumerate(self.filters):
-                y_raw[filt.name][i] = conversions.mJys_to_mag_np(mJys[k])
-                
-        self.train_X_raw = X_raw
-        self.train_y_raw = y_raw
-
-
-        # Create validation data
-        X_raw = np.empty((self.n_val_data, len(self.parameter_names)))
-        y_raw = {filt.name: np.empty((self.n_val_data, len(self.times))) for filt in self.filters}
-
-        print(f"Creating the afterglowpy validation dataset on {self.n_val_data} random points within grid.")
-        for i in tqdm.tqdm(range(self.n_val_data)):
-            X_raw[i] = [np.random.uniform(self.parameter_grid[p][0], self.parameter_grid[p][-1]) for p in self.parameter_names]
-            param_dict = dict(zip(self.parameter_names, X_raw[i]))
-            param_dict.update(self.fixed_parameters)
-            mJys = self._call_afterglowpy(param_dict, nus)
-
-            for k, filt in enumerate(self.filters):
-                y_raw[filt.name][i] = conversions.mJys_to_mag_np(mJys[k])
-
-        self.val_X_raw = X_raw
-        self.val_y_raw = y_raw
-       
-
-
-    def _read_files(self,):
-        raw_data_train = np.load(self.outdir+"/raw_data_training.npz")
-        raw_data_validation = np.load(self.outdir+'/raw_data_validation.npz')
-
-        training_y_raw = {}
-        val_y_raw = {}
-        
-        for filt in self.filters:
-            training_y_raw[filt.name] = raw_data_train[filt.name]
-            val_y_raw[filt.name] = raw_data_validation[filt.name]
-        
-        self.n_training_data = len(raw_data_train["X_raw"])
-        self.n_val_data = len(raw_data_validation["X_raw"])
-        return raw_data_train["X_raw"], training_y_raw, raw_data_validation["X_raw"], val_y_raw
-    
-    def _set_weights(self,): # TODO: dev legacy
-        weights = {filt.name: jnp.ones(self.n_training_data) for filt in self.filters}
-        
-
-        for j, p in enumerate(self.parameter_names):
-            bins = (self.parameter_grid[p][:-1] + self.parameter_grid[p][1:])/2
-            bins = [self.parameter_grid[p][0] ,*bins, self.parameter_grid[p][-1]+0.1] # the right boundary of the bins needs to be a little larger because of digitize
-            indices = np.digitize(self.train_X_raw[:,j], bins) - 1
-            for filt in self.filters:
-                weights[filt.name] *= self.weight_grids[filt.name][p][indices]
-        
-        self.weights = {filt.name: weights[filt.name]/np.sum(weights[filt.name]) for filt in self.filters}
-        
-        
-        
-    def _call_afterglowpy(self,
-                         params_dict: dict[str, Float],
-                         nus) -> Float[Array, "n_times"]:
-        """
-        Call afterglowpy to generate a single flux density output, for a given set of parameters. Note that the parameters_dict should contain all the parameters that the model requires, as well as the nu value.
-        The output will be a set of mJys.
-
-        Args:
-            Float[Array, "n_times"]: The flux density in mJys at the given times.
-        """
-        
-        # Preprocess the params_dict into the format that afterglowpy expects, which is usually called Z
-        Z = {}
-        
-        Z["jetType"]  = params_dict.get("jetType", self.jet_type)
-        Z["specType"] = params_dict.get("specType", 0)
-        Z["z"] = params_dict.get("z", 0.0)
-        Z["xi_N"] = params_dict.get("xi_N", 1.0)
-            
-        Z["E0"]        = 10 ** params_dict["log10_E0"]
-        Z["n0"]        = 10 ** params_dict["log10_n0"]
-        Z["p"]         = params_dict["p"]
-        Z["epsilon_e"] = 10 ** params_dict["log10_epsilon_e"]
-        Z["epsilon_B"] = 10 ** params_dict["log10_epsilon_B"]
-        Z["d_L"]       = 3.086e19 # fix at 10 pc, so that AB magnitude equals absolute magnitude
-        if "inclination_EM" in list(params_dict.keys()):
-            Z["thetaObs"]  = params_dict["inclination_EM"]
-        else:
-            Z["thetaObs"]  = params_dict["thetaObs"]
-
-        if self.jet_type == -1:
-             Z["thetaCore"] = params_dict["thetaCore"]
-        
-        if "thetaWing" in list(params_dict.keys()): #for Gaussian and power law jets
-             Z["thetaWing"] = params_dict["thetaWing"]
-             Z["thetaCore"] = params_dict["xCore"]*params_dict["thetaWing"]
-
-        if self.jet_type == 4:
-            Z["b"] = params_dict["b"]
-        
-        # Afterglowpy returns flux in mJys
-        tt, nunu = np.meshgrid(self._times_afterglowpy, nus)
-        mJys = grb.fluxDensity(tt, nunu, **Z)
-        return mJys
